% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{associate_web_acl}
\alias{associate_web_acl}
\title{Associates a web ACL with a resource}
\usage{
associate_web_acl(WebACLId, ResourceArn)
}
\arguments{
\item{WebACLId}{A unique identifier (ID) for the web ACL.}

\item{ResourceArn}{The ARN (Amazon Resource Name) of the resource to be protected.}
}
\description{
Associates a web ACL with a resource.
}
\examples{

}
