% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{create_regex_pattern_set}
\alias{create_regex_pattern_set}
\title{Creates a \code{RegexPatternSet}}
\usage{
create_regex_pattern_set(
  Name = "string",
  ChangeToken = "string"
)
}
\arguments{
\item{Name}{[required] A friendly name or description of the RegexPatternSet. You can't change \code{Name} after you create a \code{RegexPatternSet}.}

\item{ChangeToken}{[required] The value returned by the most recent call to GetChangeToken.}
}
\description{
Creates a \code{RegexPatternSet}. You then use UpdateRegexPatternSet to specify the regular expression (regex) pattern that you want AWS WAF to search for, such as \code{B[a@]dB[o0]t}. You can then configure AWS WAF to reject those requests.
}
\details{
To create and configure a \code{RegexPatternSet}, perform the following steps:
\enumerate{
\item Use GetChangeToken to get the change token that you provide in the \code{ChangeToken} parameter of a \code{CreateRegexPatternSet} request.
\item Submit a \code{CreateRegexPatternSet} request.
\item Use \code{GetChangeToken} to get the change token that you provide in the \code{ChangeToken} parameter of an \code{UpdateRegexPatternSet} request.
\item Submit an UpdateRegexPatternSet request to specify the string that you want AWS WAF to watch for.
}

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the \href{http://docs.aws.amazon.com/waf/latest/developerguide/}{AWS WAF Developer Guide}.
}
\examples{

}
