% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{create_web_acl}
\alias{create_web_acl}
\title{Creates a WebACL, which contains the Rules that identify the CloudFront web requests that you want to allow, block, or count}
\usage{
create_web_acl(Name, MetricName, DefaultAction, ChangeToken)
}
\arguments{
\item{Name}{[required] A friendly name or description of the WebACL. You can't change \code{Name} after you create the \code{WebACL}.}

\item{MetricName}{[required] A friendly name or description for the metrics for this \code{WebACL}. The name can contain only alphanumeric characters (A-Z, a-z, 0-9); the name can't contain whitespace. You can't change \code{MetricName} after you create the \code{WebACL}.}

\item{DefaultAction}{[required] The action that you want AWS WAF to take when a request doesn't match the criteria specified in any of the \code{Rule} objects that are associated with the \code{WebACL}.}

\item{ChangeToken}{[required] The value returned by the most recent call to GetChangeToken.}
}
\description{
Creates a \code{WebACL}, which contains the \code{Rules} that identify the CloudFront web requests that you want to allow, block, or count. AWS WAF evaluates \code{Rules} in order based on the value of \code{Priority} for each \code{Rule}.
}
\details{
You also specify a default action, either \code{ALLOW} or \code{BLOCK}. If a web request doesn't match any of the \code{Rules} in a \code{WebACL}, AWS WAF responds to the request with the default action.

To create and configure a \code{WebACL}, perform the following steps:
\enumerate{
\item Create and update the \code{ByteMatchSet} objects and other predicates that you want to include in \code{Rules}. For more information, see CreateByteMatchSet, UpdateByteMatchSet, CreateIPSet, UpdateIPSet, CreateSqlInjectionMatchSet, and UpdateSqlInjectionMatchSet.
\item Create and update the \code{Rules} that you want to include in the \code{WebACL}. For more information, see CreateRule and UpdateRule.
\item Use GetChangeToken to get the change token that you provide in the \code{ChangeToken} parameter of a \code{CreateWebACL} request.
\item Submit a \code{CreateWebACL} request.
\item Use \code{GetChangeToken} to get the change token that you provide in the \code{ChangeToken} parameter of an UpdateWebACL request.
\item Submit an UpdateWebACL request to specify the \code{Rules} that you want to include in the \code{WebACL}, to specify the default action, and to associate the \code{WebACL} with a CloudFront distribution.
}

For more information about how to use the AWS WAF API, see the \href{http://docs.aws.amazon.com/waf/latest/developerguide/}{AWS WAF Developer Guide}.
}
\section{Accepted Parameters}{
\preformatted{create_web_acl(
  Name = "string",
  MetricName = "string",
  DefaultAction = list(
    Type = "BLOCK"|"ALLOW"|"COUNT"
  ),
  ChangeToken = "string"
)
}
}

\examples{
# The following example creates a web ACL named CreateExample.
\donttest{create_web_acl(
  ChangeToken = "abcd12f2-46da-4fdb-b8d5-fbd4c466928f",
  DefaultAction = list(
    Type = "ALLOW"
  ),
  MetricName = "CreateExample",
  Name = "CreateExample"
)}

}
