% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{delete_logging_configuration}
\alias{delete_logging_configuration}
\title{Permanently deletes the LoggingConfiguration from the specified web ACL}
\usage{
delete_logging_configuration(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the web ACL from which you want to delete the LoggingConfiguration.}
}
\description{
Permanently deletes the LoggingConfiguration from the specified web ACL.
}
\section{Accepted Parameters}{
\preformatted{delete_logging_configuration(
  ResourceArn = "string"
)
}
}

