% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{delete_permission_policy}
\alias{delete_permission_policy}
\title{Permanently deletes an IAM policy from the specified RuleGroup}
\usage{
delete_permission_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{The Amazon Resource Name (ARN) of the RuleGroup from which you want to delete the policy.

The user making the request must be the owner of the RuleGroup.}
}
\description{
Permanently deletes an IAM policy from the specified RuleGroup.
}
\details{
The user making the request must be the owner of the RuleGroup.
}
\examples{

}
