% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{delete_xss_match_set}
\alias{delete_xss_match_set}
\title{Permanently deletes an XssMatchSet}
\usage{
delete_xss_match_set(XssMatchSetId, ChangeToken)
}
\arguments{
\item{XssMatchSetId}{[required] The \code{XssMatchSetId} of the XssMatchSet that you want to delete. \code{XssMatchSetId} is returned by CreateXssMatchSet and by ListXssMatchSets.}

\item{ChangeToken}{[required] The value returned by the most recent call to GetChangeToken.}
}
\description{
Permanently deletes an XssMatchSet. You can't delete an \code{XssMatchSet} if it's still used in any \code{Rules} or if it still contains any XssMatchTuple objects.
}
\details{
If you just want to remove an \code{XssMatchSet} from a \code{Rule}, use UpdateRule.

To permanently delete an \code{XssMatchSet} from AWS WAF, perform the following steps:
\enumerate{
\item Update the \code{XssMatchSet} to remove filters, if any. For more information, see UpdateXssMatchSet.
\item Use GetChangeToken to get the change token that you provide in the \code{ChangeToken} parameter of a \code{DeleteXssMatchSet} request.
\item Submit a \code{DeleteXssMatchSet} request.
}
}
\section{Accepted Parameters}{
\preformatted{delete_xss_match_set(
  XssMatchSetId = "string",
  ChangeToken = "string"
)
}
}

\examples{
# The following example deletes an XSS match set with the ID
# example1ds3t-46da-4fdb-b8d5-abc321j569j5.
\donttest{delete_xss_match_set(
  ChangeToken = "abcd12f2-46da-4fdb-b8d5-fbd4c466928f",
  XssMatchSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)}

}
