% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{disassociate_web_acl}
\alias{disassociate_web_acl}
\title{Removes a web ACL from the specified resource, either an application load balancer or Amazon API Gateway stage}
\usage{
disassociate_web_acl(
  ResourceArn = "string"
)
}
\arguments{
\item{ResourceArn}{[required] The ARN (Amazon Resource Name) of the resource from which the web ACL is being removed, either an application load balancer or Amazon API Gateway stage.

The ARN should be in one of the following formats:
\itemize{
\item For an Application Load Balancer: \code{arn:aws:elasticloadbalancing:region:account-id:loadbalancer/app/load-balancer-name/load-balancer-id region:account-id:loadbalancer/app/load-balancer-name/load-balancer-id }
\item For an Amazon API Gateway stage: \code{arn:aws:apigateway:region::/restapis/api-id/stages/stage-name region::/restapis/api-id/stages/stage-name }
}}
}
\description{
Removes a web ACL from the specified resource, either an application load balancer or Amazon API Gateway stage.
}
\examples{

}
