% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{get_change_token_status}
\alias{get_change_token_status}
\title{Returns the status of a \code{ChangeToken} that you got by calling GetChangeToken}
\usage{
get_change_token_status(
  ChangeToken = "string"
)
}
\arguments{
\item{ChangeToken}{[required] The change token for which you want to get the status. This change token was previously returned in the \code{GetChangeToken} response.}
}
\description{
Returns the status of a \code{ChangeToken} that you got by calling GetChangeToken. \code{ChangeTokenStatus} is one of the following values:
}
\details{
\itemize{
\item \code{PROVISIONED}: You requested the change token by calling \code{GetChangeToken}, but you haven't used it yet in a call to create, update, or delete an AWS WAF object.
\item \code{PENDING}: AWS WAF is propagating the create, update, or delete request to all AWS WAF servers.
\item \code{IN_SYNC}: Propagation is complete.
}
}
\examples{
# The following example returns the status of a change token with the ID
# abcd12f2-46da-4fdb-b8d5-fbd4c466928f.
get_change_token_status(
  ChangeToken = "abcd12f2-46da-4fdb-b8d5-fbd4c466928f"
)

}
