% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{get_logging_configuration}
\alias{get_logging_configuration}
\title{Returns the LoggingConfiguration for the specified web ACL}
\usage{
get_logging_configuration(ResourceArn)
}
\arguments{
\item{ResourceArn}{The Amazon Resource Name (ARN) of the web ACL for which you want to get the LoggingConfiguration.}
}
\description{
Returns the LoggingConfiguration for the specified web ACL.
}
\examples{

}
