% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{get_rule_group}
\alias{get_rule_group}
\title{Returns the RuleGroup that is specified by the \code{RuleGroupId} that you included in the \code{GetRuleGroup} request}
\usage{
get_rule_group(
  RuleGroupId = "string"
)
}
\arguments{
\item{RuleGroupId}{[required] The \code{RuleGroupId} of the RuleGroup that you want to get. \code{RuleGroupId} is returned by CreateRuleGroup and by ListRuleGroups.}
}
\description{
Returns the RuleGroup that is specified by the \code{RuleGroupId} that you included in the \code{GetRuleGroup} request.
}
\details{
To view the rules in a rule group, use ListActivatedRulesInRuleGroup.
}
\examples{

}
