% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{get_web_acl}
\alias{get_web_acl}
\title{Returns the WebACL that is specified by WebACLId}
\usage{
get_web_acl(WebACLId)
}
\arguments{
\item{WebACLId}{[required] The \code{WebACLId} of the WebACL that you want to get. \code{WebACLId} is returned by CreateWebACL and by ListWebACLs.}
}
\description{
Returns the WebACL that is specified by \code{WebACLId}.
}
\section{Accepted Parameters}{
\preformatted{get_web_acl(
  WebACLId = "string"
)
}
}

\examples{
# The following example returns the details of a web ACL with the ID
# createwebacl-1472061481310.
\donttest{get_web_acl(
  WebACLId = "createwebacl-1472061481310"
)}

}
