% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{list_geo_match_sets}
\alias{list_geo_match_sets}
\title{Returns an array of GeoMatchSetSummary objects in the response}
\usage{
list_geo_match_sets(NextMarker = NULL, Limit = NULL)
}
\arguments{
\item{NextMarker}{If you specify a value for \code{Limit} and you have more \code{GeoMatchSet}s than the value of \code{Limit}, AWS WAF returns a \code{NextMarker} value in the response that allows you to list another group of \code{GeoMatchSet} objects. For the second and subsequent \code{ListGeoMatchSets} requests, specify the value of \code{NextMarker} from the previous response to get information about another batch of \code{GeoMatchSet} objects.}

\item{Limit}{Specifies the number of \code{GeoMatchSet} objects that you want AWS WAF to return for this request. If you have more \code{GeoMatchSet} objects than the number you specify for \code{Limit}, the response includes a \code{NextMarker} value that you can use to get another batch of \code{GeoMatchSet} objects.}
}
\description{
Returns an array of GeoMatchSetSummary objects in the response.
}
\examples{

}
