% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{list_logging_configurations}
\alias{list_logging_configurations}
\title{Returns an array of LoggingConfiguration objects}
\usage{
list_logging_configurations(NextMarker = NULL, Limit = NULL)
}
\arguments{
\item{NextMarker}{If you specify a value for \code{Limit} and you have more \code{LoggingConfigurations} than the value of \code{Limit}, AWS WAF returns a \code{NextMarker} value in the response that allows you to list another group of \code{LoggingConfigurations}. For the second and subsequent \code{ListLoggingConfigurations} requests, specify the value of \code{NextMarker} from the previous response to get information about another batch of \code{ListLoggingConfigurations}.}

\item{Limit}{Specifies the number of \code{LoggingConfigurations} that you want AWS WAF to return for this request. If you have more \code{LoggingConfigurations} than the number that you specify for \code{Limit}, the response includes a \code{NextMarker} value that you can use to get another batch of \code{LoggingConfigurations}.}
}
\description{
Returns an array of LoggingConfiguration objects.
}
\examples{

}
