% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{list_resources_for_web_acl}
\alias{list_resources_for_web_acl}
\title{Returns an array of resources associated with the specified web ACL}
\usage{
list_resources_for_web_acl(
  WebACLId = "string",
  ResourceType = "APPLICATION_LOAD_BALANCER"|"API_GATEWAY"
)
}
\arguments{
\item{WebACLId}{[required] The unique identifier (ID) of the web ACL for which to list the associated resources.}

\item{ResourceType}{The type of resource to list, either an application load balancer or Amazon API Gateway.}
}
\description{
Returns an array of resources associated with the specified web ACL.
}
\examples{

}
