% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{list_rule_groups}
\alias{list_rule_groups}
\title{Returns an array of RuleGroup objects}
\usage{
list_rule_groups(NextMarker = NULL, Limit = NULL)
}
\arguments{
\item{NextMarker}{If you specify a value for \code{Limit} and you have more \code{RuleGroups} than the value of \code{Limit}, AWS WAF returns a \code{NextMarker} value in the response that allows you to list another group of \code{RuleGroups}. For the second and subsequent \code{ListRuleGroups} requests, specify the value of \code{NextMarker} from the previous response to get information about another batch of \code{RuleGroups}.}

\item{Limit}{Specifies the number of \code{RuleGroups} that you want AWS WAF to return for this request. If you have more \code{RuleGroups} than the number that you specify for \code{Limit}, the response includes a \code{NextMarker} value that you can use to get another batch of \code{RuleGroups}.}
}
\description{
Returns an array of RuleGroup objects.
}
\examples{

}
