% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{disassociate_web_acl}
\alias{disassociate_web_acl}
\title{Removes a web ACL from the specified resource}
\usage{
disassociate_web_acl(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN (Amazon Resource Name) of the resource from which the web ACL is being removed.}
}
\description{
Removes a web ACL from the specified resource.
}
\section{Accepted Parameters}{
\preformatted{disassociate_web_acl(
  ResourceArn = "string"
)
}
}

