% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.wafregional_operations.R
\name{get_web_acl_for_resource}
\alias{get_web_acl_for_resource}
\title{Returns the web ACL for the specified resource}
\usage{
get_web_acl_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{The ARN (Amazon Resource Name) of the resource for which to get the web ACL.}
}
\description{
Returns the web ACL for the specified resource.
}
\examples{

}
